% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{breakpoints1}
\alias{breakpoints1}
\title{simple function to return crossover points from an exponential}
\usage{
breakpoints1(chr_len, rate)
}
\arguments{
\item{chr_len}{the length of the chromosome in base pairs}

\item{rate}{the rate of recombination per base pair (like 1/10^6)}
}
\description{
we could swap this out for something more complicated with varying rates and
inversions fairly easily.
}
\keyword{internal}
