% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glads_funcs.R
\name{fast_dispersal}
\alias{fast_dispersal}
\title{disperal function that avoids slow abind}
\usage{
fast_dispersal(pop1, pop2, rate1to2, rate2to1)
}
\arguments{
\item{pop1}{pop struct 1}

\item{pop2}{pop struct 2}

\item{rate1to2}{rate at which indivs in pop1 migrate to pop2}

\item{rate2tp1}{rate at which indivds in pop2 migrate to pop1}
}
\value{
Returns a list of pop structs
}
\description{
blah
}
