% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summaries.R
\name{fst_at_loci}
\alias{fst_at_loci}
\title{at each of the loci compute Fst}
\usage{
fst_at_loci(P1, P2)
}
\arguments{
\item{P1}{the pop struct, indexed by indiv, locus, gene copy, for pop 1}

\item{P2}{the pop struct for pop 2}
}
\description{
uses pegas
}

