% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink-interface.R
\name{fst_via_plink}
\alias{fst_via_plink}
\title{return a data frame of Fst values given a struct}
\usage{
fst_via_plink(R, pos)
}
\arguments{
\item{R}{a list of struct1 and struct2---the structures holding the alleles
carried by the indivs}

\item{pos}{a vector giving the positions (in base pairs) of the loci}
}
\description{
blah blah
}
