% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gid_funcs.R
\name{g2p.map}
\alias{g2p.map}
\title{make the phenotype}
\usage{
g2p.map(pop.struct, bvs, n.loci.t, ve)
}
\arguments{
\item{pop.struct}{a structure holding the population members.  A 3-D array subscripted by indivs, loci, gene-copies}

\item{bvs}{array of breeding values.  an array with n.loci rows and number of columns equal to number of alleles.}

\item{n.loci.t}{#@export}
}
\description{
This maps genotype to phenotype
}

