% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glads-package.R
\docType{package}
\name{glads-package}
\alias{glads-package}
\alias{glads}
\title{Genomic Landscape of Divergence Simulation}
\description{
The package {glads} is an individual-based framework for forward demographic, genetic and genomic simulations.
The main aim of the R package is to simulate the divergence of populations further in time, elucidating genomic patterns that may be generated by a range of demographic and genetic processes.
}
\details{
The list of functions can be displayed with \code{library(help = glads)}.

The main function of the package {glads} is \code{"evolve"}, which allows the forward simulation of the genetic structure of populations under various scenarios of drift and selection.
The supported data are allelic frequencies, SNPs or sequences of DNA.
User may either simulate a theoretical dataset or directly work with their own empirical observations.
In any case, the observation should be arranged in a two-dimensional array that represents a pair of homologous chromosomes. Each element of the array is an integer defining the copy of a given allele at a given locus.
}
