% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glads_funcs.R
\name{initial.struct}
\alias{initial.struct}
\title{a function to generate the initial population structure}
\usage{
initial.struct(n.N1, n.l, n.a.l)
}
\arguments{
\item{n.N1}{Number of individuals in population 1}

\item{n.l}{Number of loci}

\item{n.a.l}{Number of alleles at each locus.  Just set as a single scalar for all the loci right now.}
}
\description{
fill in initial values
}
