% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macs-interface.R
\name{macs_command_line}
\alias{macs_command_line}
\title{prepare a macs command line to run with system}
\usage{
macs_command_line(num = 200, length = 1e+07, A = list(t = "0.0001", r =
  "0.001", h = "100"), output = "haplotypes.txt")
}
\arguments{
\item{num}{number of sequences to simulate}

\item{length}{total length of sequence in base pairs}

\item{A}{a list of arguments to pass to macs.  If you want just the flag, pass its value as "".  See the default for an idea
of what it should be.  Each element of the list should just be a string.}

\item{output}{name of the output file}
}
\description{
this just writes that command line
}
\examples{
# a small short example.  Let's say we are going to simulate a sample of 200 sequences of
# a stretch of 10 million base
# pairs and we assume that the per-base-pair-per-generation neutral mutation rate is
# 1e-08, and that the recombination rate is roughly the same (1 cM per megabase).  Then our
# scaled rates (4Ne\\mu and 4Ne\\rho), assuming an effective size of 10,000 will be 4e-04.  This is
# what that looks like:
#
# first, the parameters
A <- list(t = 4e-04, r = 4e-04, h = 100)

# then simulate it
system(macs_command_line(200, 1e7, A = A))

# then read the results into a matrix
haps <- read_macs_output("haplotypes.txt")

# check how many sequences and variants we have
dim(haps$mat)

# see how many of those variants are singletons
sum(colSums(haps$mat-1) == 1)
}

