% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glads_funcs.R
\name{mating}
\alias{mating}
\title{Mating function}
\usage{
mating(z, G)
}
\arguments{
\item{z}{A data.frame with the sex and fitness value of each individual in the population.}

\item{G}{An object of class \code{"struct"} with the genetic structure of each individual in the population.}
}
\value{
A list of two objects of class \code{"struct"}. The first element represents the genetic structure of reproductive females and the second element the genetic structure of reproductive males in the population.
}
\description{
A function to select the mating pairs
}
\details{
The function is used internally in the function newborns() in order to select the mating pairs. This function returns a three-dimensional array. Rows represent individuals and columns the different loci. Each element of the array is an integer defining the copy of a given allele at a given locus. The third dimension of the array has two layers representing a pair of homologous chromosomes.
}
\seealso{
\code{\link{newborns}} \code{\link{evolve}}
}
\keyword{internal}
