% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glads_funcs.R
\name{migrate}
\alias{migrate}
\title{Migration function}
\usage{
migrate(G1, G2, rate1to2, rate2to1)
}
\arguments{
\item{G1}{An object of class \code{"struct"} with the genetic structute of the first population.}

\item{G2}{An object of class \code{"struct"} with the genetic structute of the second population.}

\item{rate1to2}{A numerical value indicating the rate at which individuals in G1 migrate to G2}

\item{rate2to1}{A numerical value indicating the rate at which individuals in G2 migrate to G1}
}
\value{
A list of two objects of class \code{"struct"} representing the new genetic structures of populations after migration.
}
\description{
A function for the migration of individuals between populations.
}
\details{
The function is used internally in the function evolve() in order to allow the migration of individuals between populations. This function returns a list of two-dimensional arrays that represents a pair of homologous chromosomes. Rows represent individuals and columns the different loci. Each element of the array is an integer defining the copy of a given allele at a given locus.
}
\seealso{
\code{\link{evolve}}
}
\keyword{internal}
