% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glads_funcs.R
\name{phenotype}
\alias{phenotype}
\title{Phenotype function}
\usage{
phenotype(G, bvs, add.loci, sex.ratio, e.v)
}
\arguments{
\item{G}{An object of class \code{"struct"} with the genetic structure of each individual in the population.}

\item{bvs}{A matrix with the breeding values of alleles on each loci. The number of rows is equal to the number of additive loci, while the number of columns is equal to the maximum number of alleles in a locus.}

\item{add.loci}{An integer with the total number of additive loci participating in the computation of phenotypes.}

\item{sex.ratio}{A numerical value defining the sex ratio of populations}

\item{e.v}{A numerical value defining a stochastic environmental variant in the computation of phenotypes.}
}
\value{
A data.frame with rows equal to the number of individuals and two columns ('sex' and 'z')
}
\description{
A function for the computation of additive phenotypes
}
\details{
This function is used internally in the function evolve() of type 'selection' in order to compute the additive phenotype value of individuals.

This phenotype function (\eqn{z}) focuses on an additive genetic genotype-phenotype map. The sum of values of alleles at each locus gives a breeding value (\eqn{b_v}) for each individual at a given locus. The sum of breeding values \eqn{bvs} across loci gives a breeding value for the phenotypes (\eqn{z}), which is computed as follows:
\deqn{z = \sum_{v =1 }^{n_a} b_v + \varepsilon_e(0, \sigma_e)}
Where \eqn{n_a} is equal to 'add.loci' and \eqn{\sigma_e} is equal to 'e.v'. The environmental contribution \eqn{\varepsilon_e} is assumed to be stochastic and normally distributed, with a mean of 0 and standard variation 'e.v'.
}
\examples{
## We first create a random population with 100 individuals and 10 loci
N <- 100  # Population size
nl <- 10  # Number of additive loci
na <- 4  # Number of alleles per locus
G <- initial.struct(N,nl,na)

## Additional parameters are needed for the computation of phenotypes
bvs <- t(array( seq(0,1, length = na) ,c(na, nl)))
sex.ratio <- 0.5
e.v=0.01

## Now we compute the additive phenotype value of individuals
phenotype(G, bvs, nl, sex.ratio, e.v)
}
\references{
Quilodrán, C. S., Ruegg, K., Sendell-Price, A. T., Anderson, E., Coulson, T. and Clegg, S.  (2019).
The multiple population genetic and demographic routes to islands of genomic divergence.
\emph{bioRxiv}.
\doi{10.1101/673483}.
}
\seealso{
\code{\link{evolve}} \code{\link{fitness}}
}
