% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_dispersal_placement}
\alias{rcpp_dispersal_placement}
\title{dispersal function in rcpp}
\usage{
rcpp_dispersal_placement(P1, P2, d1, d2, a1, a2)
}
\arguments{
\item{P1}{first pop struct, indexed by indiv, locus, gene copy}

\item{P2}{second pop struct}

\item{d1}{dim of first pop struct}

\item{d2}{dim of second pop struct}

\item{a1}{assignments of individuals in pop 1 to either pop 1 or 2}

\item{a2}{assignments of individuals in pop 2 to either pop 1 or 2}
}
\description{
a quick rcpp based implementation because the abind in the R implementation
gobbles up a lot of time.  For this you just pass in a vector for pop1 and a vector for
pop2 that says where each individual goes.  That way we can do the individual selection
outside of this function.
}
\keyword{internal}
