% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_g2p_map}
\alias{rcpp_g2p_map}
\title{rcpp version of g2p.map to see how it fares speedwise}
\usage{
rcpp_g2p_map(G, dims, bvs, num_loci_t, v_e)
}
\arguments{
\item{G}{the structure giving the genotypes of the indviduals.  Actually a 3-D array indexed by indiv, locus, gene copy}

\item{dims}{the dimensions of the 3-D array G for internal use.}

\item{bvs}{matrix of the breeding values, indexed by loci and alleles}

\item{num_loci_t}{the number of loci, essentially something to get the mean Z to be 0}

\item{v_e}{the environmental variance in the phenotype}
}
\description{
Just a reimplementation using Rcpp.  This returns a matrix.  The first column
are the phenotype values and the second column are the sexes.
}
