% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_recombo_segregate}
\alias{rcpp_recombo_segregate}
\title{rcpp version of function that does recombination and segregation}
\usage{
rcpp_recombo_segregate(G, dims, rf)
}
\arguments{
\item{G}{the structure giving the genotypes of the indviduals.  Actually a 3-D array indexed by indiv, locus, gene copy}

\item{dims}{the dimensions of the 3-D array G for internal use.}

\item{rf}{vector of recombination fractions.  There should be one minus the number of loci. The first
one corresponds to recombination between the first and the second marker.  These are the probabilities
of a recombination between the markers during a meiosis.}
}
\value{
The return value is a long vector that can be squished into a matrix as appropriate to put it into
the genotype struct.
}
\description{
Note that this is hard-wired for diploidy.
}
\keyword{internal}
