% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_macs_output}
\alias{read_macs_output}
\title{read ms-formatted MaCS output into a matrix of haplotypes.}
\usage{
read_macs_output(Input)
}
\arguments{
\item{Input}{the path to the mac | msformatter output file to read in. Note that
tilde expansion will not occur on this path}
}
\description{
This is implemented in Rcpp for speed.  It is designed to read a single replicate coalescent
simulation in macs' msformatter format and convert the 0/1's to 1/2's and return in a big integer
matrix of haplotypes that can then be put together as desired later. (i.e. different pops and into
different diploid individuals)
}
\keyword{internal}
