% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gid_funcs.R
\name{recombine}
\alias{recombine}
\title{makes a gamete from a genotype}
\usage{
recombine(genotype)
}
\description{
the genotype is given as a matrix of alleles.  First column is the first haplotype and the second
column is the second haplotype.
#@export
}
