% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink-interface.R
\name{result2plink}
\alias{result2plink}
\title{write plink .ped and map files from simulated markers}
\usage{
result2plink(R, pos, plinkout = "plink")
}
\arguments{
\item{R}{a list of struct1 and struct2---the structures holding the alleles
carried by the indivs}

\item{pos}{a vector giving the positions (in base pairs) of the loci}

\item{plinkout}{the prefix of the file to write out.}
}
\description{
The just assumes that everything is on chromosom 1.
}

