% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summaries.R
\name{struct2hierfstat}
\alias{struct2hierfstat}
\title{convert the two population "struct"s to a data frame of loci that the hierfstat package can use}
\usage{
struct2hierfstat(P1, P2)
}
\arguments{
\item{P1}{the pop struct, indexed by indiv, locus, gene copy, for pop 1}

\item{P2}{the pop struct for pop 2}
}
\description{
This will work as long as there are not more than 9 alleles. Probably will never end
up using this as hierfstat makes a very SLOW calculation of Fst.
}
