% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summaries.R
\name{struct2pegas}
\alias{struct2pegas}
\title{convert the two population "struct"s to a data frame of loci that the pegas package can use}
\usage{
struct2pegas(P1, P2)

struct2pegas(P1, P2)
}
\arguments{
\item{P1}{the pop struct, indexed by indiv, locus, gene copy, for pop 1}

\item{P2}{the pop struct for pop 2}

\item{P1}{the pop struct, indexed by indiv, locus, gene copy, for pop 1}

\item{P2}{the pop struct for pop 2}
}
\description{
This will work as long as there are not more than 9 alleles.

This should work with any number of alleles
}
