% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_g2p_map}
\alias{rcpp_g2p_map}
\title{Phenotype function from C++}
\usage{
rcpp_g2p_map(G, dims, bvs, add_loci, sex_ratio, e_v)
}
\arguments{
\item{G}{An object of class \code{"struct"} with the genetic structure of each individual in the population.}

\item{dims}{A vector with the dimensions of the 3-D array G.}

\item{bvs}{A matrix with the breeding values, indexed by loci and alleles}

\item{add_loci}{A vector with the position of additive loci participating in the computation of phenotypes.}

\item{sex_ratio}{A numerical value defining the sex ratio of populations}

\item{e_v}{A numerical value defining the environmental variance in the computation of phenotypes.}
}
\description{
A function for the computation of additive phenotypes
}
\keyword{internal}
