% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_summaries.R
\name{struct2pegas}
\alias{struct2pegas}
\title{Conversion of class \code{struct} to class \code{loci}}
\usage{
struct2pegas(x)
}
\arguments{
\item{x}{List of objects of class \code{"struct"} with the initial genetic structure of populations.}
}
\description{
This function converts an object of class \code{"struct"} to an object of class \code{"loci"} that can be used by the package \code{pegas}
}
\examples{
## We first create a random population with 100 individuals and 10 biallelic loci
N <- 100  # Population size
nl <- 10  # Number of additive loci
na <- 2  # Number of alleles per locus
G <- initial.struct(N,nl,na)

## We convert the object of class 'struct' into 'loci'
struct2pegas(list(G))
}
